Wipe conveyances that have been changed since earlier on.  Notably the login process abuses the titler text and bypasses conveyances in the process, wiping them here will retrigger a load.
It's worth noting, character creation scripts and just about anything else that sets a KV (including characters.initialises) causes an out of band "push" of conveyances to the HUD via the region server (a "disseminate" call) which happens off the back of any/all KV updates.
This reaches the HUD prior to it completing login, and thus pushes the HUD text / titler text VERY early on in the login sequence, which the login sequence then repeatedly replaces with "Initialising character".  Since the conveyance thinks the HUD already has the right titler text it no longer communicates this value as a "delta" to the HUD.

We could simply "not" update the titler/HUD during login, could possibly bodge the conveyances, but instead we just wipe these 4 conveyance KVs and this will trigger the System interface to reappened the correct values to the final Response part of the login sequence...

Signed-off-by: Iain Price (LUNA) <git@predestined.net>