BugFix: RunCommand dropdown and prevent premature update
The dropdown for the 'command' in the RunCommand object type configuration was not being pre-populated with saved data, this caused the value to be cleared unintentionally when the Object's configuration page was viewed.

This commit resolves two related issues that caused the configuration for the RunCommand object type to be cleared unintentionally.

1.  In ObjectManagement the editObjectType method was unconditionally calling the update method on the object type.  In the case of the RunCommand objext this caused the data to be cleared on the initial page load before the form was even displayed. The fix now calls the update method if the request contains POST data

2.  In RunCommand the editForm method was not preselecting the currently configured command in the dropdown list. Now it will set the dropdown's value from the stored config