Was getting "Duplicate entry" error on rare occasions, which means the code is trying to insert a row into a table with a PK that already existed.
Traced the error to the apply method of the Effect class.

The problem lies in the fact that the DELETE and INSERT operations are not atomic. Creating a potential race condition where two concurrent requests can both pass the initial "effect alread on for longer" check, and then both attempt to insert the same record, leading to the dup key exception.

Replacing the separate DELETE and INSERT statements with an atomic REPLACE INTO solves.

(cherry picked from commit 584eec9029558b9a176a1468bee8ab3f2776fc95)